Split code in several files to improve cleanness
