+++ colors.cc
@@ -1,0 +1,8 @@
+//Macros de colores
+#define BLUE "\e[0;34m"   //POne el texto en color azul
+#define UBLUE "\e[4;34m"  //Color azul con subrayado
+#define RED "\e[0;31m"    //Color rojo
+#define GREEN "\e[0;32m"  //Color verde
+#define BGREEN "\e[1;32m" //Color verde con negrita
+#define RESET "\e[0m"     //Reset de formato de letra (color, negrita,..)
+#define CLEAR "\33[2J"    //Hace un clear pero mas bonito
